backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobot-backend"
version = "0.1.0"
description = "Backend for AudioBot using FastAPI and LangGraph"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "langgraph",
    "langgraph-checkpoint-redis",
    "langchain-core",
    "langchain-openai",
    "python-dotenv",
    "websockets",
    "langchain",
    "ollama",
    "langchain-community",
    "redis",
    "langchain-ollama",
    "langchain-groq",
    "faster-whisper",
    "torch",
    "torchaudio",
    "torchcodec",
    "coqui-tts==0.27.5",
    "edge-tts",
    "transformers>=4.57.0",
    "langsmith",
    "pdfplumber",
    "python-multipart"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
